chore: split sandbox funcs from util/helm/cmd
